trabalho final/requirements.txt
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.1
cffi==1.14.4
cssselect2==0.4.1
defusedxml==0.6.0
dj-static==0.0.6
Django==3.1.6
django-stdimage==5.1.1
html5lib==1.1
Pillow==8.1.0
progressbar2==3.53.1
psycopg2==2.8.6
pycparser==2.20
Pyphen==0.10.0
python-utils==2.5.5
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
tinycss2==1.1.0
WeasyPrint==52.2
webencodings==0.5.1
